# Apply performance patches to Sequelize 3