--- a/lib/associations/has-many.js
+++ b/lib/associations/has-many.js
@@ -272,7 +272,7 @@ HasMany.prototype.get = function(instances, options) {
     instances = undefined;
   }
 
-  options = Utils.cloneDeep(options) || {};
+  options = Object.assign({}, options);
 
   if (association.scope) {
     _.assign(where, association.scope);