--- a/lib/dialects/abstract/query-generator.js
+++ b/lib/dialects/abstract/query-generator.js
@@ -332,7 +332,7 @@ var QueryGenerator = {
 
     var query = 'INSERT<%= ignoreDuplicates %> INTO <%= table %> (<%= attributes %>) VALUES <%= tuples %><%= onConflict %><%= onDuplicateKeyUpdate %><%= returning %>;'
       , tuples = []
-      , serials = []
+      , serials = {}
       , allAttributes = []
       , onDuplicateKeyUpdate = ''
       , onConflict = '';
@@ -344,7 +344,7 @@ var QueryGenerator = {
         }
 
         if (rawAttributes[key] && rawAttributes[key].autoIncrement === true) {
-          serials.push(key);
+          serials[key] = true;
         }
       });
     });
@@ -352,7 +352,7 @@ var QueryGenerator = {
     attrValueHashes.forEach(function(attrValueHash) {
       tuples.push('(' +
         allAttributes.map(function(key) {
-          if (this._dialect.supports.bulkDefault && serials.indexOf(key) !== -1) {
+          if (this._dialect.supports.bulkDefault && serials[key] === true) {
             return attrValueHash[key] || 'DEFAULT';
           }
           return this.escape(attrValueHash[key], rawAttributes[key], { context: 'INSERT' });