--- a/lib/dialects/abstract/query.js
+++ b/lib/dialects/abstract/query.js
@@ -102,7 +102,7 @@ var groupJoinData = function(rows, includeOptions, options) {
       if ($current.includeMap[piece]) {
         includeMap[key] = $current = $current.includeMap[piece];
         if (previousPiece) {
-          previousPiece = previousPiece+'.'+piece;
+          previousPiece = `${previousPiece}.${piece}`;
         } else {
           previousPiece = piece;
         }
@@ -153,7 +153,7 @@ var groupJoinData = function(rows, includeOptions, options) {
     , getUniqueKeyAttributes = function (model) {
         var uniqueKeyAttributes = Utils._.chain(model.uniqueKeys);
         uniqueKeyAttributes = uniqueKeyAttributes
-        .result(uniqueKeyAttributes.findKey() + '.fields')
+        .result(`${uniqueKeyAttributes.findKey()}.fields`)
         .map(function(field){
           return Utils._.findKey(model.attributes, function(chr) {
             return chr.field === field;
@@ -231,22 +231,22 @@ var groupJoinData = function(rows, includeOptions, options) {
 
           if (length) {
             for (i = 0; i < length; i++) {
-              prefix = $parent ? $parent+'.'+$prevKeyPrefix[i] : $prevKeyPrefix[i];
+              prefix = $parent ? `${$parent}.${$prevKeyPrefix[i]}` : $prevKeyPrefix[i];
               primaryKeyAttributes = includeMap[prefix].model.primaryKeyAttributes;
               $length = primaryKeyAttributes.length;
               itemHash = prefix;
               if ($length === 1) {
-                itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[0]]);
+                itemHash += stringify(row[`${prefix}.${primaryKeyAttributes[0]}`]);
               }
               else if ($length > 1) {
                 for ($i = 0; $i < $length; $i++) {
-                  itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[$i]]);
+                  itemHash += stringify(row[`${prefix}.${primaryKeyAttributes[$i]}`]);
                 }
               }
               else if (!Utils._.isEmpty(includeMap[prefix].model.uniqueKeys)) {
                 uniqueKeyAttributes = getUniqueKeyAttributes(includeMap[prefix].model);
                 for ($i = 0; $i < uniqueKeyAttributes.length; $i++) {
-                  itemHash += row[prefix+'.'+uniqueKeyAttributes[$i]];
+                  itemHash += row[`${prefix}.${uniqueKeyAttributes[$i]}`];
                 }
               }
               if (!parentHash) {
@@ -318,22 +318,22 @@ var groupJoinData = function(rows, includeOptions, options) {
 
       if (length) {
         for (i = 0; i < length; i++) {
-          prefix = $parent ? $parent+'.'+$prevKeyPrefix[i] : $prevKeyPrefix[i];
+          prefix = $parent ? `${$parent}.${$prevKeyPrefix[i]}` : $prevKeyPrefix[i];
           primaryKeyAttributes = includeMap[prefix].model.primaryKeyAttributes;
           $length = primaryKeyAttributes.length;
           itemHash = prefix;
           if ($length === 1) {
-            itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[0]]);
+            itemHash += stringify(row[`${prefix}.${primaryKeyAttributes[0]}`]);
           }
           else if ($length > 0) {
             for ($i = 0; $i < $length; $i++) {
-              itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[$i]]);
+              itemHash += stringify(row[`${prefix}.${primaryKeyAttributes[$i]}`]);
             }
           }
           else if (!Utils._.isEmpty(includeMap[prefix].model.uniqueKeys)) {
             uniqueKeyAttributes = getUniqueKeyAttributes(includeMap[prefix].model);
             for ($i = 0; $i < uniqueKeyAttributes.length; $i++) {
-              itemHash += row[prefix+'.'+uniqueKeyAttributes[$i]];
+              itemHash += row[`${prefix}.${uniqueKeyAttributes[$i]}`];
             }
           }
           if (!parentHash) {