--- a/lib/dialects/abstract/query.js
+++ b/lib/dialects/abstract/query.js
@@ -167,6 +167,7 @@ var groupJoinData = function(rows, includeOptions, options) {
     , uniqueKeyAttributes
     , prefix;
 
+  const stringify = obj => obj instanceof Buffer ? obj.toString('hex') : obj;
   for (rowsI = 0; rowsI < rowsLength; rowsI++) {
     row = rows[rowsI];
 
@@ -183,11 +184,11 @@ var groupJoinData = function(rows, includeOptions, options) {
       $length = includeOptions.model.primaryKeyAttributes.length;
       topHash = '';
       if ($length === 1) {
-        topHash = row[includeOptions.model.primaryKeyAttributes[0]];
+        topHash = stringify(row[includeOptions.model.primaryKeyAttributes[0]]);
       }
       else if ($length > 1) {
         for ($i = 0; $i < $length; $i++) {
-          topHash += row[includeOptions.model.primaryKeyAttributes[$i]];
+          topHash += stringify(row[includeOptions.model.primaryKeyAttributes[$i]]);
         }
       }
       else if (!Utils._.isEmpty(includeOptions.model.uniqueKeys)) {
@@ -235,11 +236,11 @@ var groupJoinData = function(rows, includeOptions, options) {
               $length = primaryKeyAttributes.length;
               itemHash = prefix;
               if ($length === 1) {
-                itemHash += row[prefix+'.'+primaryKeyAttributes[0]];
+                itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[0]]);
               }
               else if ($length > 1) {
                 for ($i = 0; $i < $length; $i++) {
-                  itemHash += row[prefix+'.'+primaryKeyAttributes[$i]];
+                  itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[$i]]);
                 }
               }
               else if (!Utils._.isEmpty(includeMap[prefix].model.uniqueKeys)) {
@@ -322,11 +323,11 @@ var groupJoinData = function(rows, includeOptions, options) {
           $length = primaryKeyAttributes.length;
           itemHash = prefix;
           if ($length === 1) {
-            itemHash += row[prefix+'.'+primaryKeyAttributes[0]];
+            itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[0]]);
           }
           else if ($length > 0) {
             for ($i = 0; $i < $length; $i++) {
-              itemHash += row[prefix+'.'+primaryKeyAttributes[$i]];
+              itemHash += stringify(row[prefix+'.'+primaryKeyAttributes[$i]]);
             }
           }
           else if (!Utils._.isEmpty(includeMap[prefix].model.uniqueKeys)) {