--- a/lib/model.js
+++ b/lib/model.js
@@ -1351,8 +1351,7 @@ Model.prototype.findAll = function(options) {
   if (arguments.length > 1) {
     throw new Error('Please note that find* was refactored and uses only one options object from now on.');
   }
-  var tableNames = {}
-    , originalOptions;
+  var tableNames = {};
 
   tableNames[this.getTableName(options)] = true;
   options = Utils.cloneDeep(options);
@@ -1403,9 +1402,8 @@ Model.prototype.findAll = function(options) {
       return this.runHooks('beforeFindAfterOptions', options);
     }
   }).then(function() {
-    originalOptions = Utils.cloneDeep(options);
-    options.tableNames = Object.keys(tableNames);
-    return this.QueryInterface.select(this, this.getTableName(options), options);
+    const selectOptions = Object.assign({}, options, { tableNames: Object.keys(tableNames) });
+    return this.QueryInterface.select(this, this.getTableName(selectOptions), selectOptions);
   }).tap(function(results) {
     if (options.hooks) {
       return this.runHooks('afterFind', results, options);
@@ -1423,7 +1421,7 @@ Model.prototype.findAll = function(options) {
       }
     }
 
-    return Model.$findSeparate(results, originalOptions);
+    return Model.$findSeparate(results, options);
   });
 };