--- a/lib/query-interface.js
+++ b/lib/query-interface.js
@@ -662,10 +662,8 @@ QueryInterface.prototype.bulkDelete = function(tableName, identifier, options, m
   return this.sequelize.query(sql, options);
 };
 
-QueryInterface.prototype.select = function(model, tableName, options) {
-  options = Utils.cloneDeep(options);
-  options.type = QueryTypes.SELECT;
-  options.model = model;
+QueryInterface.prototype.select = function(model, tableName, optionsArg) {
+  const options = Object.assign({}, optionsArg, { type: QueryTypes.SELECT, model });
 
   return this.sequelize.query(
     this.QueryGenerator.selectQuery(tableName, options, model),