I found a few performance patches that might improve Sequelize perf. The bulkInsert one is a pretty big gain but probably not related to what we're seeing. Otherwise, there's seems to be ~10% improvement but may be more improved in production where we have some degenerate cases.

For comparison, I ran Sequelize's benchmarks:
```

Control:
Owens-MacBook-Pro:sequelize owen$ DIALECT=postgres sequelize-benchmark
  sequelize:perf Using postgres DIALECT. +0ms
(node:65740) DeprecationWarning: Using the automatically created return value from client.query as an event emitter is deprecated and will be removed in pg@7.0. Please see the upgrade guide at https://node-postgres.com/guides/upgrading
  sequelize:perf Env ready +212ms
  sequelize:perf Starting findAll +8s
  sequelize:perf finding 1500 rows x 1.73 ops/sec ±2.55% (13 runs sampled) (mean 0.577634s) +8s
  sequelize:perf finding(raw) 1500 rows x 1.82 ops/sec ±0.88% (13 runs sampled) (mean 0.550458s) +7s
  sequelize:perf finding(plain) 1500 rows x 1.79 ops/sec ±1.53% (13 runs sampled) (mean 0.557214s) +7s
  sequelize:perf finding 1500 rows with join x 1.73 ops/sec ±0.69% (13 runs sampled) (mean 0.578711s) +8s
  sequelize:perf finding 1500 rows with join (heavy includes) x 1.53 ops/sec ±1.61% (12 runs sampled) (mean 0.652015s) +8s
  sequelize:perf finding 10 rows x 124 ops/sec ±0.65% (80 runs sampled) (mean 0.008034s) +6s
  sequelize:perf finding(raw) 10 rows x 128 ops/sec ±0.59% (82 runs sampled) (mean 0.007800s) +6s
  sequelize:perf finding(plain) 10 rows x 127 ops/sec ±1.22% (81 runs sampled) (mean 0.007878s) +6s
  sequelize:perf finding 10 rows with join x 116 ops/sec ±1.14% (76 runs sampled) (mean 0.008637s) +6s
  sequelize:perf Starting bulkCreate +74ms
  sequelize:perf inserting 10000 rows x 1.21 ops/sec ±1.05% (11 runs sampled) (mean 0.824985s) +9s
  sequelize:perf inserting 1000 rows x 17.05 ops/sec ±6.33% (43 runs sampled) (mean 0.058662s) +6s
  sequelize:perf inserting 10 rows x 359 ops/sec ±3.65% (72 runs sampled) (mean 0.002789s) +6s
  sequelize:perf Benchmark finished +0ms

(All updates)

Owens-MacBook-Pro:sequelize owen$ DIALECT=postgres sequelize-benchmark
  sequelize:perf Using postgres DIALECT. +0ms
(node:67244) DeprecationWarning: Using the automatically created return value from client.query as an event emitter is deprecated and will be removed in pg@7.0. Please see the upgrade guide at https://node-postgres.com/guides/upgrading
  sequelize:perf Env ready +224ms
  sequelize:perf Starting findAll +5s
  sequelize:perf finding 1500 rows x 1.81 ops/sec ±4.62% (14 runs sampled) (mean 0.552731s) +8s
  sequelize:perf finding(raw) 1500 rows x 1.91 ops/sec ±1.34% (14 runs sampled) (mean 0.524339s) +8s
  sequelize:perf finding(plain) 1500 rows x 1.86 ops/sec ±1.22% (14 runs sampled) (mean 0.536404s) +8s
  sequelize:perf finding 1500 rows with join x 1.73 ops/sec ±1.55% (13 runs sampled) (mean 0.577722s) +8s
  sequelize:perf finding 1500 rows with join (heavy includes) x 1.59 ops/sec ±1.35% (12 runs sampled) (mean 0.628003s) +8s
  sequelize:perf finding 10 rows x 127 ops/sec ±0.97% (80 runs sampled) (mean 0.007863s) +6s
  sequelize:perf finding(raw) 10 rows x 131 ops/sec ±0.81% (82 runs sampled) (mean 0.007626s) +6s
  sequelize:perf finding(plain) 10 rows x 128 ops/sec ±0.95% (81 runs sampled) (mean 0.007783s) +6s
  sequelize:perf finding 10 rows with join x 118 ops/sec ±1.05% (76 runs sampled) (mean 0.008449s) +6s
  sequelize:perf Starting bulkCreate +93ms
  sequelize:perf inserting 10000 rows x 2.23 ops/sec ±1.65% (16 runs sampled) (mean 0.448327s) +7s
  sequelize:perf inserting 1000 rows x 20.87 ops/sec ±3.21% (52 runs sampled) (mean 0.047926s) +6s
  sequelize:perf inserting 10 rows x 323 ops/sec ±28.55% (77 runs sampled) (mean 0.003097s) +7s
  sequelize:perf Benchmark finished +0ms
```