@stevenFromConvoy 2019-09-17
We should be vocal with the rollout on this. Probably during peak usage tomorrow (and when everyone's around) would be best.